Near final update - Last headers updated - Tweaks to panic state - Added fleeing state + debug - Added option to switch off player group autonomous suppression (based on feedback) - Reduced default radio ranges to 1000 based on gameplay - Added fnc_taskAssault*
fnc_taskAssault is the replacement for fnc_taskAttack (which does nothing special).  It is a dumb rushing script as requested by Beagle.

It has however given me a bit of a headache. The current implementation is not 'ready'. Being in a dev. phase variables are not cleaned up-- nor does it take any advantage of CBA/HEMTT build. (I've just tested it in a mission file).

There are two versions. One which relies on recursive CBA task wait Untils -- the other, this is the production version, which uses cba perframe eventhandler.

The reason for the second option is because I imagine this is a waypoint which ZEUS will constantly try to update. Also; I find that units need constant impetuous to keep going.

So WIP.

-k