Fix file headers and perform minor tweaks (#43)
* HERE WE GO!
- Added new file headers to all files
- Fixed uncommented line in settings.sqf
- Tweaked hide inside range changed from 50 to 3
- Removed automatic suppression for player led groups

* Updating some wp

* - Fix Updated mod name to danger.fsm
- Fix Prevent Assault when unit is busy with certain orders
- Fix tweaking group manoeuvre actions
- Fix set all danger functions to public - no scope

* - Diwako tweak

* Near final update
- Last headers updated
- Tweaks to panic state
- Added fleeing state + debug
- Added option to switch off player group autonomous suppression (based on feedback)
- Reduced default radio ranges to 1000 based on gameplay
- Added fnc_taskAssault*

fnc_taskAssault is the replacement for fnc_taskAttack (which does nothing special).  It is a dumb rushing script as requested by Beagle.

It has however given me a bit of a headache. The current implementation is not 'ready'. Being in a dev. phase variables are not cleaned up-- nor does it take any advantage of CBA/HEMTT build. (I've just tested it in a mission file).

There are two versions. One which relies on recursive CBA task wait Untils -- the other, this is the production version, which uses cba perframe eventhandler.

The reason for the second option is because I imagine this is a waypoint which ZEUS will constantly try to update. Also; I find that units need constant impetuous to keep going.

So WIP.

-k

* - Fixes to fn_fleeing
- Adds break-out routines fn_assaultbuilding
- FSM Fix to variable and speed setting in regards to CQC mode