Improving doDodge running in place
While largely mitigated the problem still occurs.  It is my theory that it is a matter of the AI being given multiple commands from multiple sources -- that again meet the vagaries of the animation system.

In this PR I experiment with two options.

One is "_unit forceSpeed 0" when the unit will move backwards.  That has had good effect in doCover

Another is to delay the animation one frame.  Which also seems to tweak the issue, though not work perfectly.