Prevents sharing information to player group
Prevents sharing information to player led group. This has the benefit of stopping the group from entering "combat mode" outside the players control. While it does mean that player groups gain less information to engage targets at long range-- the _player_ is present. So that will rarely be an issue!