Move panic feature from brain to eventhhandlers
This pull request moves the panic feature from the brain to the eventhandler. This has the benefit of simplifying the FSM and making the panic feature more dependable.

The panic feature will not trigger for units that have a positive morale score, are not suppressed, or have their AI forced in some way.

The panic feature is enhanced by reintroducing the panic types existant in lambs danger fsm 2.4.4