Improving doDodge running in place (#228)
* Improving doDodge running in place

While largely mitigated the problem still occurs.  It is my theory that it is a matter of the AI being given multiple commands from multiple sources -- that again meet the vagaries of the animation system.

In this PR I experiment with two options.

One is "_unit forceSpeed 0" when the unit will move backwards.  That has had good effect in doCover

Another is to delay the animation one frame.  Which also seems to tweak the issue, though not work perfectly.

* Update fnc_doDodge.sqf

Annotates more of the experimental features

- Adds indoor check to prevent dodging through buildings. Does make the AI less active :/
- Reorders exit conditions to actually let disableAI dodge work!

* Update fnc_doDodge.sqf

Work in progress. Lots of debug messages. this version is working very well on 2.5.1 dev version 4.  Very little sliding around.

Exploring whether or not the "Suppress" state interferes with dodging.

* Simplify doDodge.sqf function