Fix issues relating to constant dismounts
This patches some of the issues pertaining to constant dismounting of units.

It does this by rethinking how information is shared, and fixing a bug in doFleeing (which would not take into account the units current situation very well).

Information is now only passed for units that are within "combatShareRange"  ~ 200 meters by default.  Units outside this distance, but within radio range, will unless busy, instead turn towards the direction of the unit calling for aid.

Hiding is also tweaked to make that contact response more effective, which again translates into fewer repetetive executions of said command.