Allows AI to target ACE3 incapacitated enemies in CQB
Potentially a controversial fix, but nonetheless grimly effective.

By allowing the AI to find and target ACE3 incapacitated enemies in CQB we FIX an inherent problem with the ACE3 system.

Even though the incapacitated unit is rendered invisible, the FSM on an engine level still "knows" there is something there. But because the unit is invisible, it will merely start faffing about.

By allowing the AI to simply move there and kill the downed soldier, the problem resolves itself-- and the enemy becomes all the more deadly in CQB.

Being incapacitated outside of CQB range is unchanged (and players and AI will still survive)